aws-sdk-2.2: Add test with core-only.
