+++ instrumentation/aws-sdk/aws-sdk-2.2/library/src/testCoreOnly/groovy/io/opentelemetry/instrumentation/awssdk/v2_2/Aws2ClientDynamodbTest.groovy
@@ -1,0 +1,22 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.instrumentation.awssdk.v2_2
+
+import io.opentelemetry.instrumentation.test.LibraryTestTrait
+import software.amazon.awssdk.core.client.config.ClientOverrideConfiguration
+
+class Aws2ClientTest extends AbstractAws2ClientCoreTest implements LibraryTestTrait {
+  @Override
+  ClientOverrideConfiguration.Builder createOverrideConfigurationBuilder() {
+    return ClientOverrideConfiguration.builder()
+        .addExecutionInterceptor(
+            AwsSdkTelemetry.builder(getOpenTelemetry())
+                .setCaptureExperimentalSpanAttributes(true)
+                .setUseConfiguredPropagatorForMessaging(isSqsAttributeInjectionEnabled())
+                .build()
+                .newExecutionInterceptor())
+  }
+}
